fix(equivalence/service): make unsupported variant classes explicit
Two spots silently returned None for variant classes they did not handle,
so unsupported inputs looked like a clean "no result" rather than carrying
an explicit signal. Address #810 at each site according to what its None
actually means.

equivalence/checker.rs: the Allele arm of check_accession_version_difference
is correct as-is — that helper only detects whether two variants are the
same change on different accession versions, and an allele has no single
accession version to compare (accession() reports only its first member).
None here is "not-applicable"; check() falls through to the normalizer,
which handles alleles, so allele equivalence is fully determined. Replace
the misleading "Complex case, skip for now" comment with an accurate
rationale and add regression tests proving allele pairs flow through check()
without erroring.

service/tools/ferro.rs and service/handlers/validate.rs: eliminate the
accidental silence in extract_variant_details. ferro.rs was dropping Rna,
Mt, and Circular (all carry a loc_edit and can produce real details) to
None; validate.rs was dropping Rna (Mt/Circular already covered). Add the
missing arms so these axes are populated. Replace the catch-all _ => None
with explicit named arms for the classes that have no single
coordinate/position to flatten (RnaFusion, GenomeRing, Supernumerary,
Allele, NullAllele, UnknownAllele); a future HgvsVariant arm now fails to
compile here rather than being silently dropped.

No Option/Result contract changes: converting either site to error would
regress allele checks (checker) or ripple through the JSON/web contract
(service). None remains a legitimate "no structured breakdown" value.
