+++ src/service/handlers/validate.rs
@@ -6,7 +6,7 @@ use std::time::Instant;
 use crate::hgvs::interval::interval_is_wraparound;
 use crate::service::{
     server::AppState,
-    types::{ErrorResponse, ParsedVariantDetails, PositionDetails, ServiceError, ValidateResponse},
+    types::{extract_variant_details, ErrorResponse, ServiceError, ValidateResponse},
     validation::validate_hgvs as security_validate_hgvs,
 };
 
@@ -136,364 +136,46 @@ pub fn validate_hgvs(hgvs: &str) -> ValidateResponse {
     }
 }
 
-/// Extract parsed variant details from an HgvsVariant
-fn extract_variant_details(
-    variant: &crate::hgvs::variant::HgvsVariant,
-) -> Option<ParsedVariantDetails> {
-    use crate::hgvs::variant::HgvsVariant;
-
-    match variant {
-        HgvsVariant::Cds(v) => {
-            let (variant_type, deleted, inserted) = if let Some(edit) = v.loc_edit.edit.inner() {
-                extract_na_edit_info(edit)
-            } else {
-                ("unknown".to_string(), None, None)
-            };
-            Some(ParsedVariantDetails {
-                reference: v.accession.to_string(),
-                coordinate_system: "c".to_string(),
-                variant_type,
-                position: PositionDetails {
-                    start: 0,
-                    end: None,
-                    offset: None,
-                    display: v.loc_edit.location.to_string(),
-                },
-                deleted,
-                inserted,
-                was_shifted: None,
-                original_position: None,
-            })
-        }
-        HgvsVariant::Genome(v) => {
-            let (variant_type, deleted, inserted) = if let Some(edit) = v.loc_edit.edit.inner() {
-                extract_na_edit_info(edit)
-            } else {
-                ("unknown".to_string(), None, None)
-            };
-            Some(ParsedVariantDetails {
-                reference: v.accession.to_string(),
-                coordinate_system: "g".to_string(),
-                variant_type,
-                position: PositionDetails {
-                    start: 0,
-                    end: None,
-                    offset: None,
-                    display: v.loc_edit.location.to_string(),
-                },
-                deleted,
-                inserted,
-                was_shifted: None,
-                original_position: None,
-            })
-        }
-        HgvsVariant::Tx(v) => {
-            let (variant_type, deleted, inserted) = if let Some(edit) = v.loc_edit.edit.inner() {
-                extract_na_edit_info(edit)
-            } else {
-                ("unknown".to_string(), None, None)
-            };
-            Some(ParsedVariantDetails {
-                reference: v.accession.to_string(),
-                coordinate_system: "n".to_string(),
-                variant_type,
-                position: PositionDetails {
-                    start: 0,
-                    end: None,
-                    offset: None,
-                    display: v.loc_edit.location.to_string(),
-                },
-                deleted,
-                inserted,
-                was_shifted: None,
-                original_position: None,
-            })
-        }
-        HgvsVariant::Protein(v) => Some(ParsedVariantDetails {
-            reference: v.accession.to_string(),
-            coordinate_system: "p".to_string(),
-            variant_type: "protein_change".to_string(),
-            position: PositionDetails {
-                start: 0,
-                end: None,
-                offset: None,
-                display: v.loc_edit.location.to_string(),
-            },
-            deleted: None,
-            inserted: None,
-            was_shifted: None,
-            original_position: None,
-        }),
-        HgvsVariant::Mt(v) => {
-            let (variant_type, deleted, inserted) = if let Some(edit) = v.loc_edit.edit.inner() {
-                extract_na_edit_info(edit)
-            } else {
-                ("unknown".to_string(), None, None)
-            };
-            Some(ParsedVariantDetails {
-                reference: v.accession.to_string(),
-                coordinate_system: "m".to_string(),
-                variant_type,
-                position: PositionDetails {
-                    start: 0,
-                    end: None,
-                    offset: None,
-                    display: v.loc_edit.location.to_string(),
-                },
-                deleted,
-                inserted,
-                was_shifted: None,
-                original_position: None,
-            })
-        }
-        HgvsVariant::Circular(v) => {
-            let (variant_type, deleted, inserted) = if let Some(edit) = v.loc_edit.edit.inner() {
-                extract_na_edit_info(edit)
-            } else {
-                ("unknown".to_string(), None, None)
-            };
-            Some(ParsedVariantDetails {
-                reference: v.accession.to_string(),
-                coordinate_system: "o".to_string(),
-                variant_type,
-                position: PositionDetails {
-                    start: 0,
-                    end: None,
-                    offset: None,
-                    display: v.loc_edit.location.to_string(),
-                },
-                deleted,
-                inserted,
-                was_shifted: None,
-                original_position: None,
-            })
-        }
-        _ => None,
-    }
-}
-
-/// Extract edit type and sequences from NaEdit
-fn extract_na_edit_info(
-    edit: &crate::hgvs::edit::NaEdit,
-) -> (String, Option<String>, Option<String>) {
-    use crate::hgvs::edit::NaEdit;
-
-    match edit {
-        NaEdit::Substitution {
-            reference,
-            alternative,
-        } => (
-            "substitution".to_string(),
-            Some(reference.to_string()),
-            Some(alternative.to_string()),
-        ),
-        NaEdit::SubstitutionNoRef { alternative } => (
-            "substitution".to_string(),
-            None,
-            Some(alternative.to_string()),
-        ),
-        NaEdit::Deletion { sequence, length } => {
-            let deleted = sequence
-                .as_ref()
-                .map(|s| s.to_string())
-                .or_else(|| length.map(|l| format!("{} bp", l)));
-            ("deletion".to_string(), deleted, None)
-        }
-        NaEdit::Insertion { sequence } => {
-            ("insertion".to_string(), None, Some(sequence.to_string()))
-        }
-        NaEdit::Delins {
-            sequence,
-            deleted,
-            deleted_length,
-        } => {
-            let deleted = deleted
-                .as_ref()
-                .map(|s| s.to_string())
-                .or_else(|| deleted_length.map(|l| format!("{} bp", l)));
-            ("delins".to_string(), deleted, Some(sequence.to_string()))
-        }
-        NaEdit::Duplication {
-            sequence, length, ..
-        } => {
-            let deleted = sequence
-                .as_ref()
-                .map(|s| s.to_string())
-                .or_else(|| length.map(|l| format!("{} bp", l)));
-            ("duplication".to_string(), deleted, None)
-        }
-        NaEdit::Inversion { sequence, length } => {
-            let deleted = sequence
-                .as_ref()
-                .map(|s| s.to_string())
-                .or_else(|| length.map(|l| format!("{} bp", l)));
-            ("inversion".to_string(), deleted, None)
-        }
-        NaEdit::Repeat {
-            sequence, count, ..
-        } => {
-            let seq = sequence.as_ref().map(|s| s.to_string());
-            ("repeat".to_string(), seq, Some(format!("{}", count)))
-        }
-        NaEdit::Identity { .. } => ("identity".to_string(), None, None),
-        NaEdit::Unknown { .. } => ("unknown".to_string(), None, None),
-        _ => ("other".to_string(), None, None),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
+    // Per-arm coverage of the shared `extract_variant_details` /
+    // `extract_na_edit_info` helpers lives in `crate::service::types`. Here we
+    // cover the validate handler's own surface: that `validate_hgvs` wires the
+    // shared component breakdown into the response and that `variant_wraps_origin`
+    // is axis-aware.
+
     #[test]
-    fn test_extract_variant_details_cds() {
+    fn test_validate_hgvs_populates_components() {
+        let response = validate_hgvs("NM_000249.4:c.350C>T");
+        assert!(response.valid);
+        let components = response.components.expect("components populated");
+        assert_eq!(components.coordinate_system, "c");
+        assert_eq!(components.variant_type, "substitution");
+        assert_eq!(components.reference, "NM_000249.4");
+    }
+
+    #[test]
+    fn test_validate_hgvs_allele_has_no_components() {
+        // Compound alleles have no single coordinate/position to flatten, so the
+        // shared extractor returns None and the response carries no breakdown.
+        let response = validate_hgvs("NM_000088.3:c.[10A>G;20C>T]");
+        assert!(response.valid);
+        assert!(response.components.is_none());
+    }
+
+    #[test]
+    fn test_validate_hgvs_invalid_input() {
+        let response = validate_hgvs("not a variant");
+        assert!(!response.valid);
+        assert!(response.components.is_none());
+        assert!(!response.wraps_origin);
+    }
+
+    #[test]
+    fn test_variant_wraps_origin_linear_axis_is_false() {
         let result = crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350C>T").unwrap();
-        let details = extract_variant_details(&result.result);
-
-        assert!(details.is_some());
-        let d = details.unwrap();
-        assert_eq!(d.coordinate_system, "c");
-        assert_eq!(d.variant_type, "substitution");
-        assert_eq!(d.reference, "NM_000249.4");
-    }
-
-    #[test]
-    fn test_extract_variant_details_genomic() {
-        let result =
-            crate::hgvs::parser::parse_hgvs_lenient("NC_000007.14:g.117559593G>A").unwrap();
-        let details = extract_variant_details(&result.result);
-
-        assert!(details.is_some());
-        let d = details.unwrap();
-        assert_eq!(d.coordinate_system, "g");
-        assert_eq!(d.variant_type, "substitution");
-    }
-
-    #[test]
-    fn test_extract_variant_details_protein() {
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NP_000240.1:p.Val600Glu").unwrap();
-        let details = extract_variant_details(&result.result);
-
-        assert!(details.is_some());
-        let d = details.unwrap();
-        assert_eq!(d.coordinate_system, "p");
-        assert_eq!(d.variant_type, "protein_change");
-    }
-
-    #[test]
-    fn test_extract_variant_details_mitochondrial() {
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NC_012920.1:m.8993T>G").unwrap();
-        let details = extract_variant_details(&result.result);
-
-        assert!(details.is_some());
-        let d = details.unwrap();
-        assert_eq!(d.coordinate_system, "m");
-    }
-
-    #[test]
-    fn test_extract_variant_details_noncoding() {
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NR_000001.1:n.100A>G").unwrap();
-        let details = extract_variant_details(&result.result);
-
-        assert!(details.is_some());
-        let d = details.unwrap();
-        assert_eq!(d.coordinate_system, "n");
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_substitution() {
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350C>T").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "substitution");
-                assert_eq!(deleted, Some("C".to_string()));
-                assert_eq!(inserted, Some("T".to_string()));
-            }
-        }
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_deletion() {
-        // W3025 (DelExplicitSeq) strips the explicit sequence in lenient mode,
-        // so the parsed NaEdit::Deletion has sequence=None after preprocessing.
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350delC").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "deletion");
-                assert_eq!(deleted, None);
-                assert!(inserted.is_none());
-            }
-        }
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_insertion() {
-        let result =
-            crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350_351insATG").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "insertion");
-                assert!(deleted.is_none());
-                assert_eq!(inserted, Some("ATG".to_string()));
-            }
-        }
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_delins() {
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350delinsATG").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "delins");
-                assert_eq!(deleted, None, "short form has no explicit deleted");
-                assert_eq!(inserted, Some("ATG".to_string()));
-            }
-        }
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_delins_with_explicit_deleted_seq() {
-        let result =
-            crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350_352delATGinsTTCC").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "delins");
-                assert_eq!(deleted, Some("ATG".to_string()));
-                assert_eq!(inserted, Some("TTCC".to_string()));
-            }
-        }
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_delins_with_explicit_deleted_length() {
-        let result =
-            crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350_352del3insTA").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "delins");
-                assert_eq!(deleted, Some("3 bp".to_string()));
-                assert_eq!(inserted, Some("TA".to_string()));
-            }
-        }
-    }
-
-    #[test]
-    fn test_extract_na_edit_info_duplication() {
-        // W3024 (DupExplicitSeq) strips the explicit sequence in lenient mode,
-        // so the parsed NaEdit::Duplication has sequence=None after preprocessing.
-        let result = crate::hgvs::parser::parse_hgvs_lenient("NM_000249.4:c.350dupC").unwrap();
-        if let crate::hgvs::variant::HgvsVariant::Cds(v) = &result.result {
-            if let Some(edit) = v.loc_edit.edit.inner() {
-                let (vtype, deleted, _inserted) = extract_na_edit_info(edit);
-                assert_eq!(vtype, "duplication");
-                assert_eq!(deleted, None);
-            }
-        }
+        assert!(!variant_wraps_origin(&result.result));
     }
 }
